refactor(completion): split completion tests into srp submodule (#0000)
