+++ crates/perl-lsp-rs-core/src/providers/completion/completion.rs
@@ -1737,2330 +1737,4 @@ impl CompletionProvider {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use perl_parser_core::Parser;
-    use perl_tdd_support::{must, must_some};
-    use perl_workspace::workspace_index::WorkspaceIndex;
-    use std::fs;
-    use std::path::PathBuf;
-    use std::sync::Arc;
-    use tempfile::TempDir;
-    use url::Url;
-
-    #[test]
-    fn test_variable_completion() {
-        let code = r#"
-my $count = 42;
-my $counter = 0;
-my @items = ();
-
-$c
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len() - 1);
-
-        assert!(completions.iter().any(|c| c.label == "$count"));
-        assert!(completions.iter().any(|c| c.label == "$counter"));
-    }
-
-    #[test]
-    fn test_function_completion() {
-        let code = r#"
-sub process_data {
-    # ...
-}
-
-sub process_items {
-    # ...
-}
-
-proc
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len() - 1);
-
-        assert!(completions.iter().any(|c| c.label == "process_data"));
-        assert!(completions.iter().any(|c| c.label == "process_items"));
-    }
-
-    #[test]
-    fn test_use_constant_completion_from_visible_symbol_table() {
-        let code = r#"
-package My::Config;
-use constant PI => 3.14159;
-use constant qw(MAX_RETRIES TIMEOUT);
-
-P
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len() - 1);
-
-        let pi_completion = completions.iter().find(|c| c.label == "PI");
-        assert!(pi_completion.is_some(), "expected PI constant completion");
-        assert_eq!(
-            pi_completion.map(|c| c.kind),
-            Some(crate::providers::completion_item::CompletionItemKind::Constant)
-        );
-    }
-
-    #[test]
-    fn test_use_constant_hash_form_completion() {
-        // Verify hash-ref form `use constant { FOO => 1, BAR => 2 }` surfaces both names.
-        let code = r#"
-use constant { MIN_VAL => 1, MAX_VAL => 100 };
-
-M
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len() - 1);
-
-        let min_completion = completions.iter().find(|c| c.label == "MIN_VAL");
-        assert!(
-            min_completion.is_some(),
-            "MIN_VAL should appear in completions from hash-form use constant"
-        );
-        assert_eq!(
-            min_completion.map(|c| c.kind),
-            Some(crate::providers::completion_item::CompletionItemKind::Constant),
-            "MIN_VAL should have kind Constant"
-        );
-
-        let max_completion = completions.iter().find(|c| c.label == "MAX_VAL");
-        assert!(
-            max_completion.is_some(),
-            "MAX_VAL should appear in completions from hash-form use constant"
-        );
-    }
-
-    #[test]
-    fn test_use_constant_no_parens_in_insert_text() {
-        // Constants must insert without trailing () — unlike function completions.
-        let code = r#"
-use constant ANSWER => 42;
-
-A
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len() - 1);
-
-        let answer = completions.iter().find(|c| c.label == "ANSWER");
-        assert!(answer.is_some(), "ANSWER should appear in completions");
-        assert_eq!(
-            answer.and_then(|c| c.insert_text.as_deref()),
-            Some("ANSWER"),
-            "Constants must not insert trailing () — they are called like barewords"
-        );
-    }
-
-    #[test]
-    fn test_builtin_completion() {
-        let code = "pr";
-
-        let mut parser = Parser::new(""); // Empty AST
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(completions.iter().any(|c| c.label == "print"));
-        assert!(completions.iter().any(|c| c.label == "printf"));
-    }
-
-    #[test]
-    fn test_current_package_detection() {
-        let code = r#"package Foo;
-my $x = 1;
-$x
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-
-        // position at end of file
-        let context = provider.analyze_context(code, code.len());
-        assert_eq!(context.current_package, "Foo");
-    }
-
-    #[test]
-    fn test_package_block_detection() {
-        let code = r#"package Foo {
-    my $x;
-    $x;
-}
-package Bar;
-$"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-
-        // Inside Foo block
-        let pos_foo = must_some(code.find("$x;")) + 2; // position after $x
-        let ctx_foo = provider.analyze_context(code, pos_foo);
-        assert_eq!(ctx_foo.current_package, "Foo");
-
-        // After block, in Bar package
-        let pos_bar = code.len();
-        let ctx_bar = provider.analyze_context(code, pos_bar);
-        assert_eq!(ctx_bar.current_package, "Bar");
-    }
-
-    #[test]
-    fn test_incomplete_nested_block_scope_context() {
-        let code = concat!(
-            "my $file_var = 0;\n",
-            "sub process {\n",
-            "    my $sub_var = 1;\n",
-            "    if (1) {\n",
-            "        my $block_var = 2;\n",
-            "        $"
-        );
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-        let context = provider.analyze_context(code, code.len());
-
-        let sub_scope = must_some(
-            provider
-                .symbol_table
-                .symbols
-                .get("sub_var")
-                .and_then(|symbols| symbols.first())
-                .map(|symbol| symbol.scope_id),
-        );
-        let block_scope = must_some(
-            provider
-                .symbol_table
-                .symbols
-                .get("block_var")
-                .and_then(|symbols| symbols.first())
-                .map(|symbol| symbol.scope_id),
-        );
-
-        assert_eq!(
-            context.cursor_scope_id, block_scope,
-            "expected cursor scope to match block_var scope in incomplete nested block; cursor={:?} sub={:?} block={:?}",
-            context.cursor_scope_id, sub_scope, block_scope
-        );
-    }
-
-    #[test]
-    fn test_incomplete_nested_block_variable_sorting() {
-        let code = concat!(
-            "my $file_var = 0;\n",
-            "sub process {\n",
-            "    my $sub_var = 1;\n",
-            "    if (1) {\n",
-            "        my $block_var = 2;\n",
-            "        $"
-        );
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-        let completions = provider.get_completions(code, code.len());
-
-        let block_item =
-            must_some(completions.iter().find(|completion| completion.label == "$block_var"));
-        let sub_item =
-            must_some(completions.iter().find(|completion| completion.label == "$sub_var"));
-
-        assert!(
-            block_item.sort_text < sub_item.sort_text,
-            "expected incomplete block variable to outrank parent variable, got block={:?} sub={:?}",
-            block_item.sort_text,
-            sub_item.sort_text
-        );
-    }
-
-    #[test]
-    fn test_variable_completion_prefers_nearest_parent_scope_over_name_order() {
-        let code = concat!(
-            "{\n",
-            "    my $v_a = 1;\n",
-            "    {\n",
-            "        my $v_z = 2;\n",
-            "        {\n",
-            "            $v_\n",
-            "        }\n",
-            "    }\n",
-            "}\n"
-        );
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-        // Use rfind to locate the standalone $v_ completion trigger (the LAST $v_ in
-        // the source), not the first occurrence which is inside $v_a or $v_z.
-        let trigger_pos = must_some(code.rfind("$v_")) + 3;
-        let completions = provider.get_completions(code, trigger_pos);
-
-        let v_a_idx =
-            must_some(completions.iter().position(|completion| completion.label == "$v_a"));
-        let v_z_idx =
-            must_some(completions.iter().position(|completion| completion.label == "$v_z"));
-
-        assert!(
-            v_z_idx < v_a_idx,
-            "expected one-hop parent variable ($v_z) to rank before two-hop parent ($v_a); indices: v_z={v_z_idx}, v_a={v_a_idx}"
-        );
-    }
-
-    #[test]
-    fn test_package_member_completion() {
-        // Create workspace index with a module exporting a function
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = must(Url::parse("file:///workspace/MyModule.pm"));
-        let module_code = r#"package MyModule;
-our @EXPORT = qw(exported_sub);
-sub exported_sub { }
-sub internal_sub { }
-1;
-"#;
-        must(index.index_file(module_uri, module_code.to_string()));
-
-        // Code that triggers package completion
-        let code = "use MyModule;\nMyModule::";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "exported_sub"),
-            "should suggest exported_sub"
-        );
-        let exported_sub =
-            must_some(completions.iter().find(|completion| completion.label == "exported_sub"));
-        let documentation = must_some(exported_sub.documentation.as_deref());
-        assert!(
-            documentation.contains("MyModule::exported_sub"),
-            "expected package member doc to mention qualified symbol, got: {documentation:?}"
-        );
-    }
-
-    #[test]
-    fn test_moo_accessor_method_completion() {
-        let code = r#"
-package Example::User;
-use Moo;
-
-has 'name' => (is => 'ro', isa => 'Str');
-
-sub greet {
-    my $self = shift;
-    return $self->name;
-}
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-
-        let synthesized = provider
-            .symbol_table
-            .symbols
-            .get("name")
-            .map(|symbols| symbols.iter().any(|symbol| symbol.kind == SymbolKind::Subroutine))
-            .unwrap_or(false);
-        assert!(synthesized, "expected synthesized `name` subroutine symbol in symbol table");
-
-        let pos = must_some(code.find("$self->name")) + "$self->".len();
-        let completions = provider.get_completions(code, pos);
-
-        assert!(
-            completions.iter().any(|item| item.label == "name"),
-            "expected synthesized Moo accessor `name` in method completion"
-        );
-    }
-
-    #[test]
-    fn test_moo_accessor_completion_shows_isa_type() {
-        let code = r#"
-package Example::User;
-use Moo;
-
-has 'name' => (is => 'ro', isa => 'Str');
-has 'age'  => (is => 'rw', isa => 'Int');
-
-sub greet {
-    my $self = shift;
-    $self->
-}
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-
-        let pos = must_some(code.find("$self->")) + "$self->".len();
-        let completions = provider.get_completions(code, pos);
-
-        // name accessor should appear with isa type in documentation
-        let name_item = must_some(completions.iter().find(|c| c.label == "name"));
-        let name_doc = must_some(name_item.documentation.as_deref());
-        assert!(
-            name_doc.contains("Str"),
-            "expected `Str` type in name accessor documentation, got: {name_doc:?}"
-        );
-
-        // age accessor should appear with isa type in documentation
-        let age_item = must_some(completions.iter().find(|c| c.label == "age"));
-        let age_doc = must_some(age_item.documentation.as_deref());
-        assert!(
-            age_doc.contains("Int"),
-            "expected `Int` type in age accessor documentation, got: {age_doc:?}"
-        );
-
-        // detail should indicate it's a Moo/Moose accessor, not just "method"
-        let name_detail = must_some(name_item.detail.as_deref());
-        assert!(
-            name_detail.contains("accessor"),
-            "expected 'accessor' in detail for Moo attribute, got: {name_detail:?}"
-        );
-    }
-
-    #[test]
-    fn test_moose_accessor_completion_shows_isa_type() {
-        let code = r#"
-package Example::Animal;
-use Moose;
-
-has 'species' => (is => 'ro', isa => 'Str', required => 1);
-
-sub describe {
-    my $self = shift;
-    $self->
-}
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-
-        let pos = must_some(code.find("$self->")) + "$self->".len();
-        let completions = provider.get_completions(code, pos);
-
-        let species_item = must_some(completions.iter().find(|c| c.label == "species"));
-        let species_doc = must_some(species_item.documentation.as_deref());
-        assert!(
-            species_doc.contains("Str"),
-            "expected `Str` type in species accessor documentation, got: {species_doc:?}"
-        );
-    }
-
-    #[test]
-    fn test_moo_has_option_key_completion() {
-        let code = r#"
-use Moo;
-has 'name' => (re
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|item| item.label == "required"),
-            "expected `required` option completion inside has(...) context"
-        );
-        assert!(
-            completions.iter().any(|item| item.label == "reader"),
-            "expected `reader` option completion inside has(...) context"
-        );
-    }
-
-    #[test]
-    fn test_moo_has_option_key_completion_with_quoted_prefix() {
-        let code = r#"
-use Moo;
-has 'name' => ('re
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|item| item.label == "required"),
-            "expected `required` option completion for quoted key prefix"
-        );
-    }
-
-    #[test]
-    fn test_object_pad_constructor_param_completion() {
-        let code = r#"
-use Object::Pad;
-
-class Point {
-    field $x :param = 0;
-    field $y :param = 0;
-    field $cache = 1;
-}
-
-Point->new(
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|item| item.label == "x"),
-            "expected `x` constructor completion inside Point->new(...)"
-        );
-        assert!(
-            completions.iter().any(|item| item.label == "y"),
-            "expected `y` constructor completion inside Point->new(...)"
-        );
-        assert!(
-            !completions.iter().any(|item| item.label == "cache"),
-            "non-:param fields should not appear in constructor completion"
-        );
-
-        let x_item = must_some(completions.iter().find(|item| item.label == "x"));
-        assert_eq!(x_item.insert_text.as_deref(), Some("x => "));
-    }
-
-    #[test]
-    fn test_object_pad_constructor_param_completion_honors_prefix_and_value_context() {
-        let prefix_code = r#"
-use Object::Pad;
-
-class Point {
-    field $name :param;
-    field $native_name :param;
-    field $age :param;
-}
-
-Point->new(na"#;
-
-        let mut parser = Parser::new(prefix_code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, prefix_code, None);
-        let completions = provider.get_completions(prefix_code, prefix_code.len());
-        let constructor_labels: Vec<&str> = completions
-            .iter()
-            .filter(|item| item.detail.as_deref() == Some("Object::Pad constructor parameter"))
-            .map(|item| item.label.as_str())
-            .collect();
-
-        assert!(constructor_labels.contains(&"name"), "expected `name` to match prefix `na`");
-        assert!(
-            constructor_labels.contains(&"native_name"),
-            "expected `native_name` to remain available when matching prefix"
-        );
-        assert!(
-            !constructor_labels.contains(&"age"),
-            "non-matching constructor params should be filtered by prefix"
-        );
-
-        let value_code = r#"
-use Object::Pad;
-
-class Point {
-    field $name :param;
-    field $native_name :param;
-}
-
-Point->new(name => "#;
-
-        let mut parser = Parser::new(value_code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, value_code, None);
-        let value_completions = provider.get_completions(value_code, value_code.len());
-        let value_constructor_labels: Vec<&str> = value_completions
-            .iter()
-            .filter(|item| item.detail.as_deref() == Some("Object::Pad constructor parameter"))
-            .map(|item| item.label.as_str())
-            .collect();
-
-        assert!(
-            !value_constructor_labels.contains(&"name"),
-            "constructor key completions should not appear in value position"
-        );
-        assert!(
-            !value_constructor_labels.contains(&"native_name"),
-            "constructor key completions should not appear after `=>`"
-        );
-    }
-
-    #[test]
-    fn test_object_pad_constructor_param_completion_supports_lowercase_class_names() {
-        let code = r#"
-use Object::Pad;
-
-class point {
-    field $name :param;
-    field $native_name :param;
-}
-
-point->new(na"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-        let completions = provider.get_completions(code, code.len());
-        let constructor_labels: Vec<&str> = completions
-            .iter()
-            .filter(|item| item.detail.as_deref() == Some("Object::Pad constructor parameter"))
-            .map(|item| item.label.as_str())
-            .collect();
-
-        assert!(constructor_labels.contains(&"name"));
-        assert!(constructor_labels.contains(&"native_name"));
-    }
-
-    #[test]
-    fn test_moo_isa_type_completion_includes_builtins_and_imports() {
-        let code = r#"
-use MyApp::Types qw(UserID PositiveInt);
-use Moose;
-
-has 'id' => (
-    is => 'ro',
-    isa => 
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|item| item.label == "Str"),
-            "expected built-in Moose type `Str` in isa completion, got: {:?}",
-            completions.iter().map(|item| &item.label).collect::<Vec<_>>()
-        );
-        assert!(
-            completions.iter().any(|item| item.label == "ArrayRef"),
-            "expected built-in Moose type `ArrayRef` in isa completion"
-        );
-        assert!(
-            completions.iter().any(|item| item.label == "UserID"),
-            "expected imported custom type `UserID` in isa completion"
-        );
-    }
-
-    #[test]
-    fn test_moo_isa_type_completion_with_quoted_prefix() {
-        let code = r#"
-use Moose;
-
-has 'id' => (
-    is => 'ro',
-    isa => 'St
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source(&ast, code, None);
-
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|item| item.label == "Str"),
-            "expected built-in Moose type `Str` for quoted isa prefix"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_binding_operator() {
-        // Cursor right after the opening slash of a regex
-        let code = r#"my $x = "hello"; $x =~ /"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        // Should contain regex constructs
-        assert!(
-            completions.iter().any(|c| c.label == "\\d"),
-            "expected \\d regex completion inside =~ /.../"
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "\\w"),
-            "expected \\w regex completion inside =~ /.../"
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "(?:...)"),
-            "expected non-capturing group regex completion"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_negated_binding() {
-        let code = r#"my $x = "test"; $x !~ /"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "\\d"),
-            "expected regex completions after !~"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_m_operator() {
-        let code = "if ($line =~ m/";
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "\\d"),
-            "expected regex completions inside m/.../"
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "^"),
-            "expected anchor completions inside m/.../"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_qr_operator() {
-        let code = "my $re = qr/";
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "\\d+"),
-            "expected common pattern completions inside qr/.../"
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "(?=...)"),
-            "expected lookahead group completion inside qr/.../"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_s_operator() {
-        let code = "($line = $input) =~ s/";
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "\\s+"),
-            "expected common pattern completions inside s/.../"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_has_all_categories() {
-        let code = r#"$x =~ /"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        // Character classes
-        assert!(completions.iter().any(|c| c.label == "\\d"));
-        assert!(completions.iter().any(|c| c.label == "\\D"));
-        assert!(completions.iter().any(|c| c.label == "\\w"));
-        assert!(completions.iter().any(|c| c.label == "\\W"));
-        assert!(completions.iter().any(|c| c.label == "\\s"));
-        assert!(completions.iter().any(|c| c.label == "\\S"));
-        assert!(completions.iter().any(|c| c.label == "\\h"));
-        assert!(completions.iter().any(|c| c.label == "\\H"));
-        assert!(completions.iter().any(|c| c.label == "\\v"));
-        assert!(completions.iter().any(|c| c.label == "\\V"));
-        assert!(completions.iter().any(|c| c.label == "\\R"));
-        assert!(completions.iter().any(|c| c.label == "[...]"));
-        assert!(completions.iter().any(|c| c.label == "[^...]"));
-
-        // Anchors
-        assert!(completions.iter().any(|c| c.label == "^"));
-        assert!(completions.iter().any(|c| c.label == "$"));
-        assert!(completions.iter().any(|c| c.label == "\\b"));
-        assert!(completions.iter().any(|c| c.label == "\\B"));
-        assert!(completions.iter().any(|c| c.label == "\\A"));
-        assert!(completions.iter().any(|c| c.label == "\\z"));
-        assert!(completions.iter().any(|c| c.label == "\\Z"));
-
-        // Quantifiers
-        assert!(completions.iter().any(|c| c.label == "*"));
-        assert!(completions.iter().any(|c| c.label == "+"));
-        assert!(completions.iter().any(|c| c.label == "?"));
-        assert!(completions.iter().any(|c| c.label == "{n}"));
-        assert!(completions.iter().any(|c| c.label == "{n,}"));
-        assert!(completions.iter().any(|c| c.label == "{n,m}"));
-
-        // Groups
-        assert!(completions.iter().any(|c| c.label == "(...)"));
-        assert!(completions.iter().any(|c| c.label == "(?:...)"));
-        assert!(completions.iter().any(|c| c.label == "(?=...)"));
-        assert!(completions.iter().any(|c| c.label == "(?!...)"));
-        assert!(completions.iter().any(|c| c.label == "(?<=...)"));
-        assert!(completions.iter().any(|c| c.label == "(?<!...)"));
-
-        // Common patterns
-        assert!(completions.iter().any(|c| c.label == "\\d+"));
-        assert!(completions.iter().any(|c| c.label == "\\w+"));
-        assert!(completions.iter().any(|c| c.label == "\\s+"));
-        assert!(completions.iter().any(|c| c.label == ".*?"));
-        assert!(completions.iter().any(|c| c.label == ".+?"));
-    }
-
-    #[test]
-    fn test_regex_completion_items_have_correct_kind() {
-        let code = r#"$x =~ /"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        for item in &completions {
-            assert_eq!(
-                item.kind,
-                CompletionItemKind::Snippet,
-                "regex completion '{}' should be Snippet kind",
-                item.label
-            );
-        }
-    }
-
-    #[test]
-    fn test_regex_completion_items_have_documentation() {
-        let code = r#"$x =~ /"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        for item in &completions {
-            assert!(
-                item.documentation.is_some(),
-                "regex completion '{}' should have documentation",
-                item.label
-            );
-            assert!(item.detail.is_some(), "regex completion '{}' should have detail", item.label);
-        }
-    }
-
-    #[test]
-    fn test_regex_completion_not_in_normal_context() {
-        // Outside regex context, should not get regex completions
-        let code = "my $x = 1;\n";
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            !completions.iter().any(|c| c.label == "\\d"),
-            "regex completions should NOT appear outside regex context"
-        );
-    }
-
-    #[test]
-    fn test_is_in_regex_binding_operator() {
-        let code = r#"$x =~ /hello"#;
-        assert!(CompletionProvider::is_in_regex(code, code.len()));
-    }
-
-    #[test]
-    fn test_is_in_regex_m_operator() {
-        let code = "m/pattern";
-        assert!(CompletionProvider::is_in_regex(code, code.len()));
-    }
-
-    #[test]
-    fn test_is_in_regex_qr_operator() {
-        let code = "my $re = qr/pattern";
-        assert!(CompletionProvider::is_in_regex(code, code.len()));
-    }
-
-    #[test]
-    fn test_is_in_regex_s_operator() {
-        let code = "$line =~ s/old";
-        assert!(CompletionProvider::is_in_regex(code, code.len()));
-    }
-
-    #[test]
-    fn test_is_in_regex_keyword_operator() {
-        let code = "$x or /pattern";
-        assert!(CompletionProvider::is_in_regex(code, code.len()));
-    }
-
-    #[test]
-    fn test_is_not_in_regex_division() {
-        // Division should NOT be detected as regex
-        let code = "my $result = $x / $y";
-        // Position after "$x / $" -- should not be regex because $ precedes /
-        // but our heuristic checks pre_slash context
-        assert!(
-            !CompletionProvider::is_in_regex(code, code.len()),
-            "division should not be detected as regex context"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_preserves_sigil_completions_in_interpolation() {
-        // Cursor is inside the regex body at the end of `$fo` — before the
-        // closing `/`. Variable completions must be offered, not flag completions.
-        let code = r#"my $foo = 1; my $bar = qr/^$fo/"#;
-        // Position just before the closing '/'
-        let pos = code.len() - 1;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, pos);
-
-        assert!(
-            completions.iter().any(|item| item.label == "$foo"),
-            "expected interpolated regex variables to keep scalar completions"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_replaces_escape_prefix_range() {
-        let code = r#"$x =~ /\d"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        let item = must_some(completions.iter().find(|completion| completion.label == r"\d"));
-        assert_eq!(
-            item.text_edit_range,
-            Some((code.len() - r"\d".len(), code.len())),
-            "expected regex completion to replace the typed escape sequence"
-        );
-    }
-
-    #[test]
-    fn test_regex_completion_offers_perl_whitespace_and_linebreak_classes() {
-        let code = r#"$x =~ /\"#;
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        let labels: Vec<&str> = completions.iter().map(|c| c.label.as_str()).collect();
-
-        for label in &["\\h", "\\H", "\\v", "\\V", "\\R"] {
-            assert!(
-                labels.contains(label),
-                "expected Perl regex class completion '{label}', got: {labels:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_regex_completion_replaces_group_prefix_range() {
-        let code = r#"$x =~ /(?: "#;
-        let code = code.trim_end();
-
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        let item = must_some(completions.iter().find(|completion| completion.label == "(?:...)"));
-        assert_eq!(
-            item.text_edit_range,
-            Some((code.len() - "(?:".len(), code.len())),
-            "expected regex completion to replace the typed group opener"
-        );
-    }
-
-    #[test]
-    fn test_detect_use_qw_import_context_basic() {
-        // Cursor right after opening paren in qw()
-        let code = "use MyModule qw(";
-        let result = CompletionProvider::detect_use_qw_import_context(code, code.len());
-        assert!(result.is_some(), "should detect qw() import context");
-        let (module, prefix) =
-            result.as_ref().map(|(m, p)| (m.as_str(), p.as_str())).unwrap_or_default();
-        assert_eq!(module, "MyModule");
-        assert_eq!(prefix, "");
-    }
-
-    #[test]
-    fn test_detect_use_qw_import_context_with_prefix() {
-        let code = "use File::Basename qw(bas";
-        let result = CompletionProvider::detect_use_qw_import_context(code, code.len());
-        assert!(result.is_some(), "should detect qw() import context with prefix");
-        let (module, prefix) =
-            result.as_ref().map(|(m, p)| (m.as_str(), p.as_str())).unwrap_or_default();
-        assert_eq!(module, "File::Basename");
-        assert_eq!(prefix, "bas");
-    }
-
-    #[test]
-    fn test_detect_use_qw_import_context_with_existing_imports() {
-        let code = "use MyModule qw(foo bar ba";
-        let result = CompletionProvider::detect_use_qw_import_context(code, code.len());
-        assert!(result.is_some(), "should detect qw() import context after existing imports");
-        let (module, prefix) =
-            result.as_ref().map(|(m, p)| (m.as_str(), p.as_str())).unwrap_or_default();
-        assert_eq!(module, "MyModule");
-        assert_eq!(prefix, "ba");
-    }
-
-    #[test]
-    fn test_detect_use_qw_not_after_close() {
-        // Cursor after the closing paren
-        let code = "use MyModule qw(foo bar);";
-        let result = CompletionProvider::detect_use_qw_import_context(code, code.len());
-        assert!(result.is_none(), "should not detect context after closing paren");
-    }
-
-    #[test]
-    fn test_detect_use_qw_not_for_pragmas() {
-        let code = "use strict qw(";
-        let result = CompletionProvider::detect_use_qw_import_context(code, code.len());
-        assert!(result.is_none(), "should not detect context for lowercase pragmas");
-    }
-
-    #[test]
-    fn test_use_qw_import_completion_with_workspace() -> Result<(), Box<dyn std::error::Error>> {
-        // Create workspace index with a module that has subroutines
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyUtils.pm")?;
-        let module_code = r#"package MyUtils;
-use Exporter 'import';
-our @EXPORT_OK = qw(helper_one helper_two);
-sub helper_one { }
-sub helper_two { }
-sub _private_internal { }
-1;
-"#;
-        index.index_file(module_uri, module_code.to_string())?;
-
-        // Code where user is typing inside qw()
-        let code = "use MyUtils qw(hel";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "helper_one"),
-            "should suggest helper_one from MyUtils: got {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "helper_two"),
-            "should suggest helper_two from MyUtils"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_qw_import_completion_empty_prefix() -> Result<(), Box<dyn std::error::Error>> {
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/Utils.pm")?;
-        let module_code = r#"package Utils;
-sub alpha { }
-sub beta { }
-1;
-"#;
-        index.index_file(module_uri, module_code.to_string())?;
-
-        // Empty prefix inside qw()
-        let code = "use Utils qw(";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "alpha"),
-            "should suggest alpha with empty prefix: got {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "beta"),
-            "should suggest beta with empty prefix"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_qw_import_completion_detail_shows_module() -> Result<(), Box<dyn std::error::Error>>
-    {
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyLib.pm")?;
-        let module_code = r#"package MyLib;
-sub do_work { }
-1;
-"#;
-        index.index_file(module_uri, module_code.to_string())?;
-
-        let code = "use MyLib qw(do";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-
-        let do_work = completions.iter().find(|c| c.label == "do_work");
-        assert!(do_work.is_some(), "should suggest do_work");
-        let detail = must_some(do_work.and_then(|c| c.detail.as_deref()));
-        assert!(detail.contains("MyLib"), "detail should mention module name, got: {detail:?}");
-        Ok(())
-    }
-
-    #[test]
-    fn test_self_arrow_resolves_workspace_methods() -> Result<(), Box<dyn std::error::Error>> {
-        // Regression test for issue #2536: $self-> method completion should resolve
-        // workspace-indexed methods from the current package.
-        //
-        // The methods are ONLY in the workspace index (a separate .pm file), not in
-        // the currently-parsed source. This tests the workspace path specifically:
-        // `infer_receiver_package` must return `MyService` for `$self->` when
-        // `context.current_package == "MyService"`.
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyService.pm")?;
-        let module_code = r#"package MyService;
-sub new { bless {}, shift }
-sub process_request { }
-sub validate_input { }
-1;
-"#;
-        index.index_file(module_uri, module_code.to_string())?;
-
-        // The currently-edited file is in MyService but does NOT define
-        // process_request or validate_input locally — they are workspace-only.
-        let code = r#"package MyService;
-sub run {
-    my $self = shift;
-    $self->"#;
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "process_request"),
-            "$self-> should suggest process_request from workspace index; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "validate_input"),
-            "$self-> should suggest validate_input from workspace index"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_this_arrow_resolves_workspace_methods() -> Result<(), Box<dyn std::error::Error>> {
-        // Same as above but using $this as the invocant variable.
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyHandler.pm")?;
-        let module_code = r#"package MyHandler;
-sub new { bless {}, shift }
-sub handle { }
-1;
-"#;
-        index.index_file(module_uri, module_code.to_string())?;
-
-        // Only `run` is in the edited file; `handle` lives only in the workspace index.
-        let code = r#"package MyHandler;
-sub run {
-    my $this = shift;
-    $this->"#;
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-
-        assert!(
-            completions.iter().any(|c| c.label == "handle"),
-            "$this-> should suggest handle from workspace index; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_self_arrow_in_main_package_does_not_resolve() -> Result<(), Box<dyn std::error::Error>>
-    {
-        // Edge case: $self-> in the main package should NOT resolve to any package methods.
-        // The guard condition `context.current_package != "main"` prevents incorrect
-        // suggestions when the user is in script-level code.
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyLib.pm")?;
-        let module_code = r#"package MyLib;
-sub new { bless {}, shift }
-sub helper { }
-1;
-"#;
-        index.index_file(module_uri, module_code.to_string())?;
-
-        // Code is at package main (implicit), so $self-> should not resolve
-        let code = r#"sub run {
-    my $self = shift;
-    $self->"#;
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-
-        // Should NOT suggest MyLib methods just because the variable is named $self
-        assert!(
-            !completions.iter().any(|c| c.label == "helper"),
-            "$self-> in main package should not suggest methods from other packages"
-        );
-        Ok(())
-    }
-
-    // -------------------------------------------------------------------------
-    // Tests for is_use_statement_context and add_use_module_completions
-    // -------------------------------------------------------------------------
-
-    #[test]
-    fn test_use_statement_context_after_use_keyword() -> Result<(), Box<dyn std::error::Error>> {
-        // "use " with cursor right after space — empty prefix, should trigger module completion
-        let index = Arc::new(WorkspaceIndex::new());
-        let uri = Url::parse("file:///lib/MyApp.pm")?;
-        index.index_file(uri, "package MyApp;\n1;\n".to_string())?;
-        let code = "use ";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "MyApp" && c.kind == CompletionItemKind::Module),
-            "use <cursor> should suggest workspace module names; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_statement_context_with_prefix() -> Result<(), Box<dyn std::error::Error>> {
-        // "use MyA" — prefix filtering should narrow to MyApp, not OtherLib
-        let index = Arc::new(WorkspaceIndex::new());
-        index
-            .index_file(Url::parse("file:///lib/MyApp.pm")?, "package MyApp;\n1;\n".to_string())?;
-        index.index_file(
-            Url::parse("file:///lib/OtherLib.pm")?,
-            "package OtherLib;\n1;\n".to_string(),
-        )?;
-        let code = "use MyA";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "MyApp" && c.kind == CompletionItemKind::Module),
-            "use MyA should suggest MyApp with Module kind"
-        );
-        assert!(
-            !completions.iter().any(|c| c.label == "OtherLib"),
-            "use MyA should not suggest OtherLib"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_statement_skips_pragmas() -> Result<(), Box<dyn std::error::Error>> {
-        // Lowercase-first token after `use` means pragma — no module completion.
-        // The index is populated with a Module-kind package so that if the lowercase
-        // guard in is_use_statement_context were absent the test would fail (not
-        // vacuously pass due to an empty index).
-        let index = Arc::new(WorkspaceIndex::new());
-        index.index_file(
-            Url::parse("file:///lib/Strict.pm")?,
-            "package Strict;\n1;\n".to_string(),
-        )?;
-        let code = "use strict";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            !completions.iter().any(|c| c.kind == CompletionItemKind::Module),
-            "use strict should not trigger module completions; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_statement_skips_past_module_name_at_qw() -> Result<(), Box<dyn std::error::Error>> {
-        // Cursor inside qw list should NOT trigger module-name completion.
-        // The index is populated so the test is non-vacuous: if the qw-dispatch
-        // branch were removed, add_use_module_completions would fire and the
-        // Module-kind assertion would fail.
-        let index = Arc::new(WorkspaceIndex::new());
-        index.index_file(
-            Url::parse("file:///lib/Module.pm")?,
-            "package Module;\nsub foo {}\n1;\n".to_string(),
-        )?;
-        let code = "use Module qw(foo";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        // This context routes to qw-import completions (Function kind), not module-name
-        // completions (Module kind).
-        assert!(
-            !completions.iter().any(|c| c.kind == CompletionItemKind::Module),
-            "cursor inside qw() should not get module-name completions; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_require_statement_triggers_module_completion() -> Result<(), Box<dyn std::error::Error>>
-    {
-        let index = Arc::new(WorkspaceIndex::new());
-        index
-            .index_file(Url::parse("file:///lib/Utils.pm")?, "package Utils;\n1;\n".to_string())?;
-        let code = "require Ut";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "Utils" && c.kind == CompletionItemKind::Module),
-            "require Ut should suggest Utils with Module kind; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_require_statement_skips_file_path() -> Result<(), Box<dyn std::error::Error>> {
-        let index = Arc::new(WorkspaceIndex::new());
-        index
-            .index_file(Url::parse("file:///lib/Utils.pm")?, "package Utils;\n1;\n".to_string())?;
-        let code = "require './utils.pl'";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            !completions.iter().any(|c| c.kind == CompletionItemKind::Module),
-            "require './utils.pl' should not trigger module-name completions; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_require_statement_skips_version_check() -> Result<(), Box<dyn std::error::Error>> {
-        let index = Arc::new(WorkspaceIndex::new());
-        index
-            .index_file(Url::parse("file:///lib/Utils.pm")?, "package Utils;\n1;\n".to_string())?;
-        let code = "require 5.010";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            !completions.iter().any(|c| c.kind == CompletionItemKind::Module),
-            "require 5.010 should not trigger module-name completions; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_require_statement_triggers_completion_for_lowercase_module()
-    -> Result<(), Box<dyn std::error::Error>> {
-        // `require autodie` is valid Perl — lowercase module names must still get completions.
-        // The previous implementation incorrectly blocked all non-uppercase-starting require
-        // targets, including valid lowercase modules like autodie, overload, and Carp.
-        let index = Arc::new(WorkspaceIndex::new());
-        index.index_file(
-            Url::parse("file:///lib/autodie.pm")?,
-            "package autodie;\n1;\n".to_string(),
-        )?;
-        let code = "require auto";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions
-                .iter()
-                .any(|c| c.label == "autodie" && c.kind == CompletionItemKind::Module),
-            "require auto should suggest 'autodie' with Module kind; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_require_statement_skips_vstring_version() -> Result<(), Box<dyn std::error::Error>> {
-        // `require v5.10` is a v-string version check, not a module name.
-        // 'v' starts the token but it is not followed by '::' — it should be blocked
-        // because 'v' is a digit-prefix indicator in this context.
-        // Currently, 'v' is a letter so it passes the digit/quote/path check.
-        // This is an inherent limitation of single-char prefix detection — the full
-        // `require v5.10` case requires position-aware parsing to resolve correctly.
-        // For now, assert the observed (not-yet-blocked) behavior to document it.
-        let index = Arc::new(WorkspaceIndex::new());
-        index
-            .index_file(Url::parse("file:///lib/Utils.pm")?, "package Utils;\n1;\n".to_string())?;
-        let code = "require v5.10";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        // v5.10 is an unlikely prefix for a module (no CPAN modules start with 'v' in practice),
-        // and even if triggered, the module index has no matching 'v*' entry.
-        // Assert we never suggest Utils for this context.
-        assert!(
-            !completions.iter().any(|c| c.label == "Utils" && c.kind == CompletionItemKind::Module),
-            "require v5.10 should not suggest unrelated modules; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_module_deduplication() -> Result<(), Box<dyn std::error::Error>> {
-        // Two files declaring the same package should produce one completion, not two
-        let index = Arc::new(WorkspaceIndex::new());
-        index
-            .index_file(Url::parse("file:///lib/MyApp.pm")?, "package MyApp;\n1;\n".to_string())?;
-        index.index_file(
-            Url::parse("file:///lib/MyApp2.pm")?,
-            "package MyApp;\n1;\n".to_string(), // duplicate package name
-        )?;
-        let code = "use MyA";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        let myapp_count = completions.iter().filter(|c| c.label == "MyApp").count();
-        assert_eq!(
-            myapp_count, 1,
-            "Duplicate package declarations should produce exactly one completion"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_module_non_use_context_excluded() -> Result<(), Box<dyn std::error::Error>> {
-        // Outside a use/require statement, module-priority sort_text should NOT appear.
-        // add_use_module_completions gates on in_use_statement; its "1_" sort_text
-        // prefix is the marker we check here.
-        let index = Arc::new(WorkspaceIndex::new());
-        index.index_file(
-            Url::parse("file:///lib/MyApp.pm")?,
-            "package MyApp;\nsub hello {}\n1;\n".to_string(),
-        )?;
-        let code = "my $x = MyA";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        // The "1_MyApp" sort_text is only emitted by add_use_module_completions,
-        // which is guarded by in_use_statement. It must not appear outside that context.
-        assert!(
-            !completions.iter().any(|c| c.sort_text.as_deref() == Some("1_MyApp")),
-            "Module-priority sort_text should only appear in use context"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_path_to_module_name_maps_nested_pm_file() -> Result<(), Box<dyn std::error::Error>> {
-        let temp = TempDir::new()?;
-        let root = temp.path().join("lib");
-        let module_file = root.join("File").join("Path").join("To").join("Module.pm");
-        fs::create_dir_all(module_file.parent().ok_or("missing parent")?)?;
-        fs::write(&module_file, "package File::Path::To::Module;\n1;\n")?;
-
-        let module_name = workspace::path_to_module_name(&root, &module_file);
-        assert_eq!(module_name.as_deref(), Some("File::Path::To::Module"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_completion_scans_include_paths() -> Result<(), Box<dyn std::error::Error>> {
-        let temp = TempDir::new()?;
-        let include_root = temp.path().join("external");
-        let module_file = include_root.join("DB").join("Driver.pm");
-        fs::create_dir_all(module_file.parent().ok_or("missing parent")?)?;
-        fs::write(module_file, "package DB::Driver;\n1;\n")?;
-
-        let code = "use DB";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source_and_paths(
-            &ast,
-            code,
-            Some(Arc::new(WorkspaceIndex::new())),
-            vec![include_root],
-            Vec::new(),
-            false,
-        );
-        let completions = provider.get_completions(code, code.len());
-        assert!(completions.iter().any(|c| c.label == "DB::Driver"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_completion_workspace_first_and_dedupes_external()
-    -> Result<(), Box<dyn std::error::Error>> {
-        let temp = TempDir::new()?;
-        let include_root = temp.path().join("external");
-        let module_file = include_root.join("DBI.pm");
-        fs::create_dir_all(module_file.parent().ok_or("missing parent")?)?;
-        fs::write(module_file, "package DBI;\n1;\n")?;
-
-        let index = Arc::new(WorkspaceIndex::new());
-        index
-            .index_file(Url::parse("file:///workspace/lib/DBI.pm")?, "package DBI;\n1;\n".into())?;
-
-        let code = "use DB";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source_and_paths(
-            &ast,
-            code,
-            Some(index),
-            vec![include_root],
-            Vec::new(),
-            false,
-        );
-        let completions = provider.get_completions(code, code.len());
-        let dbi_items: Vec<_> = completions.iter().filter(|c| c.label == "DBI").collect();
-        assert_eq!(dbi_items.len(), 1, "DBI should be deduplicated across workspace/external");
-        assert_eq!(dbi_items[0].detail.as_deref(), Some("module"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_completion_system_inc_opt_in() -> Result<(), Box<dyn std::error::Error>> {
-        let temp = TempDir::new()?;
-        let system_root = temp.path().join("sys");
-        let module_file = system_root.join("Sys").join("Only.pm");
-        fs::create_dir_all(module_file.parent().ok_or("missing parent")?)?;
-        fs::write(module_file, "package Sys::Only;\n1;\n")?;
-
-        let code = "use Sys::O";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let disabled = CompletionProvider::new_with_index_and_source_and_paths(
-            &ast,
-            code,
-            Some(Arc::new(WorkspaceIndex::new())),
-            Vec::<PathBuf>::new(),
-            vec![system_root.clone()],
-            false,
-        )
-        .get_completions(code, code.len());
-        assert!(!disabled.iter().any(|c| c.label == "Sys::Only"));
-
-        let enabled = CompletionProvider::new_with_index_and_source_and_paths(
-            &ast,
-            code,
-            Some(Arc::new(WorkspaceIndex::new())),
-            Vec::<PathBuf>::new(),
-            vec![system_root],
-            true,
-        )
-        .get_completions(code, code.len());
-        let sys_only =
-            enabled.iter().find(|c| c.label == "Sys::Only").ok_or("missing Sys::Only")?;
-        assert_eq!(sys_only.detail.as_deref(), Some("system module"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_path_to_module_name_top_level_pm() -> Result<(), Box<dyn std::error::Error>> {
-        // DBI.pm directly in root → "DBI" (single-component name, no "::")
-        let temp = TempDir::new()?;
-        let root = temp.path().join("lib");
-        fs::create_dir_all(&root)?;
-        let module_file = root.join("DBI.pm");
-        fs::write(&module_file, "package DBI;\n1;\n")?;
-
-        let module_name = workspace::path_to_module_name(&root, &module_file);
-        assert_eq!(module_name.as_deref(), Some("DBI"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_path_to_module_name_non_pm_excluded() -> Result<(), Box<dyn std::error::Error>> {
-        // Only .pm files should be mapped; .pl, .pod, .so, no-ext must return None
-        let temp = TempDir::new()?;
-        let root = temp.path().join("lib");
-        fs::create_dir_all(&root)?;
-
-        for name in &["Script.pl", "Manual.pod", "XSHelper.so", "README"] {
-            let file = root.join(name);
-            fs::write(&file, "")?;
-            assert!(
-                workspace::path_to_module_name(&root, &file).is_none(),
-                "expected None for {}",
-                name
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_scan_excludes_non_pm_files() -> Result<(), Box<dyn std::error::Error>> {
-        // scan_directory_for_modules must not surface .pl or .pod files
-        let temp = TempDir::new()?;
-        let root = temp.path().join("lib");
-        fs::create_dir_all(&root)?;
-        fs::write(root.join("Good.pm"), "package Good;\n1;\n")?;
-        fs::write(root.join("Bad.pl"), "#!/usr/bin/perl\n")?;
-        fs::write(root.join("Doc.pod"), "=head1 NAME\n")?;
-
-        let modules = workspace::scan_directory_for_modules(&root, "");
-        assert!(modules.contains(&"Good".to_string()), "Good.pm should be included");
-        assert!(!modules.contains(&"Bad".to_string()), "Bad.pl should be excluded");
-        assert!(!modules.contains(&"Doc".to_string()), "Doc.pod should be excluded");
-        Ok(())
-    }
-
-    #[test]
-    fn test_scan_nonexistent_root_returns_empty() {
-        // A path that does not exist must silently return an empty list
-        let modules = workspace::scan_directory_for_modules(
-            std::path::Path::new("/nonexistent/path/that/cannot/exist/12345"),
-            "Module",
-        );
-        assert!(modules.is_empty());
-    }
-
-    #[test]
-    fn test_path_to_module_name_rejects_non_pm_extension() -> Result<(), Box<dyn std::error::Error>>
-    {
-        // path_to_module_name must return None for files that are not .pm files,
-        // even if they look like module paths.  Without this guard the scanner
-        // would surface .pl scripts and .pod documentation as completable modules.
-        let temp = TempDir::new()?;
-        let root = temp.path().join("lib");
-        fs::create_dir_all(&root)?;
-        for name in &["Script.pl", "Doc.pod", "Archive.pm.bak", "README"] {
-            let file = root.join(name);
-            fs::write(&file, "")?;
-            assert!(
-                workspace::path_to_module_name(&root, &file).is_none(),
-                "expected None for {}",
-                name
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_scan_respects_max_depth() -> Result<(), Box<dyn std::error::Error>> {
-        // scan_directory_for_modules must not descend more than MAX_SCAN_DEPTH (8)
-        // levels below the root.  A directory exactly at depth 8 should be scanned;
-        // a directory at depth 9 should be silently skipped.
-        let temp = TempDir::new()?;
-        let root = temp.path().join("lib");
-
-        // Build a path 9 levels deep: root/a/b/c/d/e/f/g/h/
-        let deep_dir =
-            root.join("a").join("b").join("c").join("d").join("e").join("f").join("g").join("h"); // depth 8 — should be scanned
-        let too_deep = deep_dir.join("x"); // depth 9 — must be skipped
-        fs::create_dir_all(&too_deep)?;
-        fs::write(deep_dir.join("AtLimit.pm"), "package A::B::C::D::E::F::G::H::AtLimit;\n1;\n")?;
-        fs::write(too_deep.join("TooDeep.pm"), "package TooDeep;\n1;\n")?;
-
-        let modules = workspace::scan_directory_for_modules(&root, "");
-        assert!(
-            modules.iter().any(|m| m == "a::b::c::d::e::f::g::h::AtLimit"),
-            "depth-8 module should be found; got: {:?}",
-            modules
-        );
-        assert!(
-            !modules.iter().any(|m| m == "x::TooDeep" || m == "TooDeep"),
-            "depth-9 module must not appear; got: {:?}",
-            modules
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_completion_not_triggered_outside_use_statement()
-    -> Result<(), Box<dyn std::error::Error>> {
-        // Verify that the scan does NOT fire for general variable completion.
-        // If a large include_root were passed and the scan ran unconditionally,
-        // this test would be slow and surface module names for non-use positions.
-        let temp = TempDir::new()?;
-        let include_root = temp.path().join("external");
-        let module_file = include_root.join("SomeMod.pm");
-        fs::create_dir_all(&include_root)?;
-        fs::write(module_file, "package SomeMod;\n1;\n")?;
-
-        // Code that does NOT contain a `use` statement — cursor at a scalar
-        let code = "my $x = 1;\nprint $x";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index_and_source_and_paths(
-            &ast,
-            code,
-            Some(Arc::new(WorkspaceIndex::new())),
-            vec![include_root],
-            Vec::new(),
-            false,
-        );
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            !completions.iter().any(|c| c.label == "SomeMod"),
-            "external module should not appear outside `use` context"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_use_statement_past_semicolon_excluded() -> Result<(), Box<dyn std::error::Error>> {
-        // Cursor at the end of `use Module;` — the semicolon guard in
-        // is_use_statement_context must suppress module-name completions.
-        // Without the `;` check the cursor would be considered still inside
-        // the use statement and would show stale module suggestions.
-        let index = Arc::new(WorkspaceIndex::new());
-        index.index_file(
-            Url::parse("file:///lib/Module.pm")?,
-            "package Module;\n1;\n".to_string(),
-        )?;
-        let code = "use Module;";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            !completions.iter().any(|c| c.kind == CompletionItemKind::Module),
-            "cursor after `use Module;` should not trigger module-name completions; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    // ── Gap 1: Named capture group in regex_patterns ─────────────────────────
-
-    #[test]
-    fn test_regex_named_capture_completion() {
-        // Cursor inside an empty regex body — named capture should be offered.
-        let code = r#"$x =~ /"#;
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "(?<name>...)"),
-            "expected named capture group in regex completions; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_regex_named_capture_prefix_disambig() {
-        // Typing `(?<` inside a regex → both lookbehind and named capture offered.
-        let code = r#"$x =~ /(?<"#;
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "(?<=...)"),
-            "expected lookbehind when prefix is (?<"
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "(?<name>...)"),
-            "expected named capture when prefix is (?<"
-        );
-    }
-
-    #[test]
-    fn test_regex_named_capture_prefix_lookbehind_only() {
-        // Typing `(?<=` — only the lookbehind should match (named capture label
-        // does not start with `(?<=`).
-        let code = r#"$x =~ /(?<="#;
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "(?<=...)"),
-            "expected lookbehind for prefix (?<="
-        );
-        assert!(
-            !completions.iter().any(|c| c.label == "(?<name>...)"),
-            "named capture should NOT appear for prefix (?<= (label doesn't start with (?<=)"
-        );
-    }
-
-    // ── Gap 2: is_in_regex_flags heuristic ───────────────────────────────────
-
-    #[test]
-    fn test_is_in_regex_flags_after_close_slash() {
-        // Cursor immediately after the closing `/` of a regex.
-        let code = "$x =~ /foo/";
-        assert!(
-            CompletionProvider::is_in_regex_flags(code, code.len()),
-            "cursor right after closing / should be in regex-flags context"
-        );
-    }
-
-    #[test]
-    fn test_is_in_regex_flags_after_partial_flag() {
-        // Cursor after one already-typed flag character.
-        let code = "m/foo/i";
-        assert!(
-            CompletionProvider::is_in_regex_flags(code, code.len()),
-            "cursor after /i should still be in regex-flags context"
-        );
-    }
-
-    #[test]
-    fn test_is_in_regex_flags_s_operator() {
-        let code = "s/foo/bar/g";
-        assert!(
-            CompletionProvider::is_in_regex_flags(code, code.len()),
-            "s/// with /g flag should be in regex-flags context"
-        );
-    }
-
-    #[test]
-    fn test_is_not_in_regex_flags_division() {
-        // Plain division — must not be treated as regex flags.
-        let code = "my $x = $a / $b /";
-        assert!(
-            !CompletionProvider::is_in_regex_flags(code, code.len()),
-            "division should not be detected as regex-flags context"
-        );
-    }
-
-    #[test]
-    fn test_regex_flag_completions_after_close() {
-        // Cursor right after closing `/` — should offer all standard flag letters.
-        let code = "$x =~ /foo/";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        let labels: Vec<&str> = completions.iter().map(|c| c.label.as_str()).collect();
-        // Standard regex flags per Perl documentation
-        for flag in &["g", "i", "m", "s", "x", "e", "r", "a", "p"] {
-            assert!(
-                labels.contains(flag),
-                "expected standard regex flag '{flag}' in completions; got: {labels:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_regex_flag_completions_skip_already_typed() {
-        // `g` already typed — completions should include `i` but not `g`.
-        let code = "$x =~ /foo/g";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        let labels: Vec<&str> = completions.iter().map(|c| c.label.as_str()).collect();
-        assert!(!labels.contains(&"g"), "already-typed flag 'g' should be excluded");
-        assert!(labels.contains(&"i"), "flag 'i' should still be offered");
-    }
-
-    #[test]
-    fn test_regex_tr_flag_completions() {
-        // tr/// should offer only c, d, s — not g, i, e.
-        let code = "tr/a-z/A-Z/";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        let labels: Vec<&str> = completions.iter().map(|c| c.label.as_str()).collect();
-        for flag in &["c", "d", "s"] {
-            assert!(
-                labels.contains(flag),
-                "tr/// flag '{flag}' should be offered; got: {labels:?}"
-            );
-        }
-        for flag in &["g", "i", "e"] {
-            assert!(!labels.contains(flag), "tr/// should NOT offer '{flag}'; got: {labels:?}");
-        }
-    }
-
-    #[test]
-    fn test_regex_tr_binding_operator_flag_completions() {
-        // `$x =~ tr/.../` should also offer only c, d, s (binding form).
-        let code = "$x =~ tr/a-z/A-Z/";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        let labels: Vec<&str> = completions.iter().map(|c| c.label.as_str()).collect();
-        for flag in &["c", "d", "s"] {
-            assert!(
-                labels.contains(flag),
-                "tr/// binding flag '{flag}' should be offered; got: {labels:?}"
-            );
-        }
-        assert!(!labels.contains(&"g"), "tr/// should NOT offer 'g'; got: {labels:?}");
-    }
-
-    // ── Gap 3: Statement-level regex operator snippets ───────────────────────
-
-    #[test]
-    fn test_regex_operator_snippets_present() {
-        let code = "";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, 0);
-        let labels: Vec<&str> = completions.iter().map(|c| c.label.as_str()).collect();
-        assert!(labels.contains(&"mregex"), "mregex snippet missing; got: {labels:?}");
-        assert!(labels.contains(&"ssubst"), "ssubst snippet missing; got: {labels:?}");
-        assert!(labels.contains(&"qrpat"), "qrpat snippet missing; got: {labels:?}");
-    }
-
-    #[test]
-    fn test_regex_operator_snippet_bodies() {
-        // Verify the insert_text for each new snippet is syntactically correct.
-        let code = "mregex";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-
-        let mregex = must_some(completions.iter().find(|c| c.label == "mregex"));
-        let insert = mregex.insert_text.as_deref().unwrap_or_default();
-        assert!(insert.starts_with("m/"), "mregex body must start with m/; got: {insert:?}");
-
-        // Also verify ssubst and qrpat with explicit prefix lookup
-        let code2 = "ssubst";
-        let mut parser2 = Parser::new(code2);
-        let ast2 = must(parser2.parse());
-        let provider2 = CompletionProvider::new(&ast2);
-        let completions2 = provider2.get_completions(code2, code2.len());
-        let ssubst = must_some(completions2.iter().find(|c| c.label == "ssubst"));
-        let insert2 = ssubst.insert_text.as_deref().unwrap_or_default();
-        assert!(insert2.starts_with("s/"), "ssubst body must start with s/; got: {insert2:?}");
-
-        let code3 = "qrpat";
-        let mut parser3 = Parser::new(code3);
-        let ast3 = must(parser3.parse());
-        let provider3 = CompletionProvider::new(&ast3);
-        let completions3 = provider3.get_completions(code3, code3.len());
-        let qrpat = must_some(completions3.iter().find(|c| c.label == "qrpat"));
-        let insert3 = qrpat.insert_text.as_deref().unwrap_or_default();
-        assert!(insert3.starts_with("qr/"), "qrpat body must start with qr/; got: {insert3:?}");
-    }
-
-    // ── Dash trigger character tests (#2865) ─────────────────────────────────
-    // When `-` is a trigger character, context detection must distinguish
-    // method-call arrows (`->`) from arithmetic/decrement operators.
-
-    #[test]
-    fn test_dash_trigger_fires_method_completion_for_arrow()
-    -> Result<(), Box<dyn std::error::Error>> {
-        // `$obj-` (cursor after `-`) — the `-` is the start of `->`, so method
-        // completions must appear even before the `>` is typed.
-        // Crucially, the result must be ONLY method completions (Function kind),
-        // not the entire keyword/snippet list. Without the `-` trigger feature,
-        // the code returns all completions — this assertion catches that false pass.
-        let code = r#"package MyService;
-sub new { bless {}, shift }
-sub process { }
-sub validate { }
-sub run {
-    my $self = shift;
-    $self-"#;
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyService.pm")?;
-        let module_code = "package MyService;\nsub process { }\nsub validate { }\n1;\n";
-        index.index_file(module_uri, module_code.to_string())?;
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        // Must find method completions from the workspace index
-        assert!(
-            completions.iter().any(|c| c.label == "process" || c.label == "validate"),
-            "dash trigger on `$self-` should produce method completions; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        // Must NOT return the full keyword/snippet dump — only method completions.
-        // "arrayref", "hashref" are snippets from the generic path; they should not
-        // appear when the context is a method-call arrow.
-        assert!(
-            !completions.iter().any(|c| c.label == "arrayref" || c.label == "hashref"),
-            "dash trigger on `$self-` must not return generic snippets; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_dash_trigger_suppressed_for_subtract_assign() {
-        // `$x -=` (cursor after `-` in `-=`) — must return NO completions.
-        let code = "$x -";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        // Position is at len() which puts cursor right after `-` preceded by space.
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.is_empty(),
-            "dash trigger on `$x -` (subtract context) should return no completions; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_dash_trigger_suppressed_for_decrement() {
-        // `$x--` — second `-` is preceded by another `-`, must return NO completions.
-        // The guard `source[position-2] != b'-'` prevents treating `--` as `->`.
-        let code = "$x--";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        // Cursor after the second `-`: preceding char is `-`, not an identifier.
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.is_empty(),
-            "dash trigger on `$x--` (decrement context) should return no completions; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_dash_trigger_suppressed_for_unary_minus() {
-        // `my $x = -$y` — unary minus, `-` preceded by space → no completions.
-        let code = "my $x = -";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.is_empty(),
-            "dash trigger on `my $x = -` (unary minus) should return no completions; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_dash_trigger_fires_for_hash_deref_arrow() -> Result<(), Box<dyn std::error::Error>> {
-        // `$hash->{key}` — trigger on `-` in `$hash->`, receiver ends with `h`
-        // (alphanumeric), should produce method completions (not a generic dump).
-        let code = r#"package MyService;
-sub new { bless {}, shift }
-sub get_data { }
-sub run {
-    my $hash = {};
-    $hash-"#;
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyService.pm")?;
-        let module_code = "package MyService;\nsub new { }\nsub get_data { }\n1;\n";
-        index.index_file(module_uri, module_code.to_string())?;
-        let provider = CompletionProvider::new_with_index(&ast, Some(index));
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "get_data" || c.label == "new"),
-            "dash trigger on `$hash-` should produce completions; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        // Must not return generic snippet dump
-        assert!(
-            !completions.iter().any(|c| c.label == "arrayref" || c.label == "hashref"),
-            "dash trigger on `$hash-` must not return generic snippets; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        Ok(())
-    }
-
-    // ── Hash key completion tests ─────────────────────────────────────────────
-
-    #[test]
-    fn test_hash_key_completion_basic() {
-        // my %config = (host => 'localhost', port => 5432);
-        // $config{ho<cursor>
-        // Expected: "host" suggested, "port" filtered out by prefix
-        let code = "my %config = (host => 'localhost', port => 5432);\n$config{ho";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "host"),
-            "expected 'host' in hash key completions for prefix 'ho'; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        assert!(
-            !completions.iter().any(|c| c.label == "port"),
-            "expected 'port' filtered out by prefix 'ho'; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_hash_key_completion_empty_prefix() {
-        // $config{<cursor> -- all keys returned
-        let code = "my %config = (host => 'localhost', port => 5432);\n$config{";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            completions.iter().any(|c| c.label == "host"),
-            "expected 'host' in hash key completions with empty prefix; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-        assert!(
-            completions.iter().any(|c| c.label == "port"),
-            "expected 'port' in hash key completions with empty prefix; got: {:?}",
-            completions.iter().map(|c| &c.label).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_hash_key_completion_does_not_fire_for_hashref_deref() {
-        // $ref->{ho<cursor> -- hashref deref, must NOT suggest hash keys
-        let code = "my $ref = {host => 'localhost'};\n$ref->{ho";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        // Must not return a Property-kinded "host" completion (hash key detection
-        // must bail when `->` precedes the `{`)
-        assert!(
-            !completions
-                .iter()
-                .any(|c| c.label == "host" && c.kind == CompletionItemKind::Property),
-            "hashref deref `$ref->{{ho` must not produce Property-kinded 'host' completion; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_hash_key_completion_in_comment_no_suggestions() {
-        // # $config{ho<cursor> -- in comment, should not suggest hash keys
-        let code = "my %config = (host => 'localhost');\n# $config{ho";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        assert!(
-            !completions
-                .iter()
-                .any(|c| c.label == "host" && c.kind == CompletionItemKind::Property),
-            "hash key completion must not fire inside a comment; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_hash_key_completion_unknown_variable_returns_empty_for_that_hash() {
-        // $config{<cursor> where %config has no known init -- no leaked keys from %other
-        let code = "my %other = (a => 1);\n$config{";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-        let provider = CompletionProvider::new(&ast);
-        let completions = provider.get_completions(code, code.len());
-        // Keys from %other must not appear as completions for $config{}
-        assert!(
-            !completions.iter().any(|c| c.label == "a" && c.kind == CompletionItemKind::Property),
-            "keys from %%other must not leak into %%config completions; got: {:?}",
-            completions.iter().map(|c| (&c.label, &c.kind)).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_detect_hash_key_context_unicode_non_ident_after_brace_no_panic() {
-        let source = "$config{☃ho";
-        let result = CompletionProvider::detect_hash_key_context(source, source.len());
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_detect_hash_key_context_unicode_non_ident_before_var_no_panic() {
-        let source = "☃config{ho";
-        let result = CompletionProvider::detect_hash_key_context(source, source.len());
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_detect_hash_key_context_unicode_before_brace_no_panic() {
-        // ☃ (3 bytes) immediately before `{` — previously caused a panic in the
-        // `->` check which sliced source[brace_pos-2..brace_pos] across a
-        // non-char-boundary.
-        let source = "$config☃{key";
-        let result = CompletionProvider::detect_hash_key_context(source, source.len());
-        // ☃ is not a valid Perl identifier char so the variable name scan will not
-        // find a `$` sigil — result must be None without panicking.
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_detect_hash_key_context_4byte_emoji_in_key_no_panic() {
-        // 4-byte emoji (U+1F600) mid-key-prefix — exercises the char_indices rev path
-        // with a surrogate-range codepoint.
-        let source = "$config{\u{1F600}ho";
-        let result = CompletionProvider::detect_hash_key_context(source, source.len());
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_detect_hash_key_context_ascii_regression() {
-        // Plain ASCII must still work correctly after the Unicode fixes.
-        let source = "$config{key";
-        let result = CompletionProvider::detect_hash_key_context(source, source.len());
-        assert_eq!(result, Some(("config".to_string(), "key".to_string())));
-    }
-
-    #[test]
-    fn test_provider_captures_include_and_system_inc_paths() {
-        let code = "use My::Module;\n";
-        let mut parser = Parser::new(code);
-        let ast = must(parser.parse());
-
-        let include_paths = vec![PathBuf::from("/workspace/lib"), PathBuf::from("t/lib")];
-        let system_inc_paths = vec![PathBuf::from("/usr/lib/perl5")];
-
-        let provider = CompletionProvider::new_with_index_and_source_and_paths(
-            &ast,
-            code,
-            None,
-            include_paths.clone(),
-            system_inc_paths.clone(),
-            false,
-        );
-
-        assert_eq!(provider.include_paths, include_paths);
-        assert_eq!(provider.system_inc_paths, system_inc_paths);
-    }
-
-    #[test]
-    fn test_use_module_completion_unchanged_with_empty_inc_vectors()
-    -> Result<(), Box<dyn std::error::Error>> {
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyApp.pm")?;
-        index.index_file(module_uri, "package MyApp;\n1;\n".to_string())?;
-
-        let code = "use MyA";
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let baseline_provider =
-            CompletionProvider::new_with_index_and_source(&ast, code, Some(Arc::clone(&index)));
-        let baseline = baseline_provider.get_completions_with_path(code, code.len(), None);
-
-        let with_empty_inc = CompletionProvider::new_with_index_and_source_and_paths(
-            &ast,
-            code,
-            Some(index),
-            Vec::new(),
-            Vec::new(),
-            false,
-        );
-        let with_empty_inc_results =
-            with_empty_inc.get_completions_with_path(code, code.len(), None);
-
-        let baseline_labels: std::collections::HashSet<String> =
-            baseline.into_iter().map(|item| item.label).collect();
-        let with_empty_labels: std::collections::HashSet<String> =
-            with_empty_inc_results.into_iter().map(|item| item.label).collect();
-
-        assert_eq!(
-            baseline_labels, with_empty_labels,
-            "empty include paths must not change completion results in phase 1"
-        );
-        Ok(())
-    }
-
-    /// Phase 1 contract: non-empty inc_paths are stored but do NOT alter completions
-    /// (no filesystem scanning until phase 2). If this test breaks it means someone
-    /// started using the paths without adding scan logic — a regression, not a feature.
-    #[test]
-    fn test_non_empty_inc_paths_do_not_change_phase1_completions()
-    -> Result<(), Box<dyn std::error::Error>> {
-        let index = Arc::new(WorkspaceIndex::new());
-        let module_uri = Url::parse("file:///workspace/MyApp.pm")?;
-        index.index_file(module_uri, "package MyApp;\n1;\n".to_string())?;
-
-        let code = "use MyA";
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let baseline =
-            CompletionProvider::new_with_index_and_source(&ast, code, Some(Arc::clone(&index)))
-                .get_completions_with_path(code, code.len(), None);
-
-        // Non-empty inc paths: completions must still be identical to workspace-only baseline.
-        let with_inc = CompletionProvider::new_with_index_and_source_and_paths(
-            &ast,
-            code,
-            Some(index),
-            vec![PathBuf::from("/usr/local/lib/perl5"), PathBuf::from("t/lib")],
-            vec![PathBuf::from("/usr/lib/perl5/5.38")],
-            false,
-        )
-        .get_completions_with_path(code, code.len(), None);
-
-        let baseline_labels: std::collections::HashSet<String> =
-            baseline.into_iter().map(|item| item.label).collect();
-        let with_inc_labels: std::collections::HashSet<String> =
-            with_inc.into_iter().map(|item| item.label).collect();
-
-        assert_eq!(
-            baseline_labels, with_inc_labels,
-            "non-empty include paths must not change completions in phase 1 (no filesystem scanning yet)"
-        );
-        Ok(())
-    }
-}
+mod tests;
